zopfli: Avoid double to float or int conversion.
No point using doubles if they're going to be treated as floats anyway.
Generally speaking, these values are only non-integral by 1/8 intervals if
I understand it correctly.

+++ zopfli/src/zopfli/tree.c
@@ -68,21 +68,21 @@ void ZopfliLengthsToSymbols(const unsigned* lengths, size_t n, unsigned maxbits,
   free(next_code);
 }
 
-void ZopfliCalculateEntropy(const size_t* count, size_t n, double* bitlengths) {
-  static const double kInvLog2 = 1.4426950408889;  /* 1.0 / log(2.0) */
+void ZopfliCalculateEntropy(const size_t* count, size_t n, float* bitlengths) {
+  static const float kInvLog2 = 1.4426950408889f;  /* 1.0 / log(2.0) */
   unsigned sum = 0;
   unsigned i;
-  double log2sum;
+  float log2sum;
   for (i = 0; i < n; ++i) {
     sum += count[i];
   }
-  log2sum = (sum == 0 ? log(n) : log(sum)) * kInvLog2;
+  log2sum = (sum == 0 ? logf(n) : logf(sum)) * kInvLog2;
   for (i = 0; i < n; ++i) {
     /* When the count of the symbol is 0, but its cost is requested anyway, it
     means the symbol will appear at least once anyway, so give it the cost as if
     its count is 1.*/
     if (count[i] == 0) bitlengths[i] = log2sum;
-    else bitlengths[i] = log2sum - log(count[i]) * kInvLog2;
+    else bitlengths[i] = log2sum - logf(count[i]) * kInvLog2;
     /* Depending on compiler and architecture, the above subtraction of two
     floating point numbers may give a negative result very close to zero
     instead of zero (e.g. -5.973954e-17 with gcc 4.1.2 on Ubuntu 11.4). Clamp
